requirements.txt
flask
gunicorn
python-dotenv
requests
reportlab
PyPDF2
python-docx
google-generativeai
firebase-admin
